Add possibility for custom event layouts
