+++ agendacalendarview/src/main/java/com/github/tibolte/agendacalendarview/CalendarPickerController.java
@@ -1,5 +1,6 @@
 package com.github.tibolte.agendacalendarview;
 
+import com.github.tibolte.agendacalendarview.models.BaseCalendarEvent;
 import com.github.tibolte.agendacalendarview.models.CalendarEvent;
 import com.github.tibolte.agendacalendarview.models.DayItem;
