+++ agendacalendarview/src/main/java/com/github/tibolte/agendacalendarview/agenda/AgendaListView.java
@@ -1,19 +1,12 @@
 package com.github.tibolte.agendacalendarview.agenda;
 
 import com.github.tibolte.agendacalendarview.CalendarManager;
-import com.github.tibolte.agendacalendarview.R;
+import com.github.tibolte.agendacalendarview.models.BaseCalendarEvent;
 import com.github.tibolte.agendacalendarview.models.CalendarEvent;
-import com.github.tibolte.agendacalendarview.utils.BusProvider;
 import com.github.tibolte.agendacalendarview.utils.DateHelper;
-import com.github.tibolte.agendacalendarview.utils.Events;
 
-import android.animation.Animator;
-import android.animation.ObjectAnimator;
 import android.content.Context;
 import android.util.AttributeSet;
-import android.view.MotionEvent;
-import android.view.ViewGroup;
-import android.view.ViewTreeObserver;
 
 import java.util.Calendar;
 import java.util.List;
