+++ agendacalendarview/src/main/res/layout/view_agenda_event.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.github.tibolte.agendacalendarview.agenda.AgendaEventView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:gravity="center_vertical"
@@ -50,4 +50,4 @@
 
     </LinearLayout>
 
-</com.github.tibolte.agendacalendarview.agenda.AgendaEventView>
+</LinearLayout>
