+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'me.tatarka.retrolambda'
 
 android {
     compileSdkVersion 23
-    buildToolsVersion "23.0.0"
+    buildToolsVersion "23.0.1"
 
     defaultConfig {
         applicationId "com.github.tibolte.sample"
@@ -43,7 +43,7 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
 
     // Google libraries
-    compile 'com.android.support:appcompat-v7:23.1.0'
+    compile 'com.android.support:appcompat-v7:23.1.1'
 
     // Other libraries
     compile 'com.jakewharton:butterknife:7.0.0'
