+++ app/src/main/java/com/github/tibolte/sample/DrawableCalendarEvent.java
@@ -1,0 +1,38 @@
+package com.github.tibolte.sample;
+
+import com.github.tibolte.agendacalendarview.models.BaseCalendarEvent;
+import com.github.tibolte.agendacalendarview.models.CalendarEvent;
+
+import java.util.Calendar;
+
+public class DrawableCalendarEvent extends BaseCalendarEvent {
+    private int drawableId;
+
+    public DrawableCalendarEvent(long id, int color, String title, String description, String location, long dateStart, long dateEnd, int allDay, String duration, int drawableId) {
+        super(id, color, title, description, location, dateStart, dateEnd, allDay, duration);
+        this.drawableId = drawableId;
+    }
+
+    public DrawableCalendarEvent(String title, String description, String location, int color, Calendar startTime, Calendar endTime, boolean allDay, int drawableId) {
+        super(title, description, location, color, startTime, endTime, allDay);
+        this.drawableId = drawableId;
+    }
+
+    public DrawableCalendarEvent(DrawableCalendarEvent calendarEvent) {
+        super(calendarEvent);
+        this.drawableId = calendarEvent.getDrawableId();
+    }
+
+    public int getDrawableId() {
+        return drawableId;
+    }
+
+    public void setDrawableId(int drawableId) {
+        this.drawableId = drawableId;
+    }
+
+    @Override
+    public CalendarEvent copy() {
+        return new DrawableCalendarEvent(this);
+    }
+}
